Move applyDecision to StateManager
